+++ src/main/java/io/github/scru128/Scru128Generator.java
@@ -1,6 +1,7 @@
 package io.github.scru128;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.security.SecureRandom;
 import java.util.Iterator;
@@ -9,24 +10,45 @@ import java.util.Random;
 
 /**
  * Represents a SCRU128 ID generator that encapsulates the monotonic counters and other internal states.
+ * <p>
+ * The generator offers four different methods to generate a SCRU128 ID:
+ * <table border="1">
+ *   <caption>Comparison of generator functions</caption>
+ *   <tr><th>Flavor</th>                       <th>Timestamp</th><th>Thread-</th><th>On big clock rewind</th></tr>
+ *   <tr><td>{@link #generate}</td>            <td>Now</td>      <td>Safe</td>   <td>Rewinds state</td></tr>
+ *   <tr><td>{@link #generateNoRewind}</td>    <td>Now</td>      <td>Safe</td>   <td>Returns null</td></tr>
+ *   <tr><td>{@link #generateCore}</td>        <td>Argument</td> <td>Unsafe</td> <td>Rewinds state</td></tr>
+ *   <tr><td>{@link #generateCoreNoRewind}</td><td>Argument</td> <td>Unsafe</td> <td>Returns null</td></tr>
+ * </table>
+ * <p>
+ * Each method returns monotonically increasing IDs unless a timestamp provided is significantly (by ten seconds or
+ * more by default) smaller than the one embedded in the immediately preceding ID. If such a significant clock
+ * rollback is detected, the {@link #generate} method rewinds the generator state and returns a new ID based on the
+ * current timestamp, whereas the experimental NoRewind variants keep the state untouched and return null. Core
+ * functions offer low-level thread-unsafe primitives.
  */
 public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<@NotNull Scru128Id> {
+    /**
+     * The default timestamp rollback allowance.
+     */
+    static final long DEFAULT_ROLLBACK_ALLOWANCE = 10000L; // 10 seconds
+
     private long timestamp = 0;
     private int counterHi = 0;
     private int counterLo = 0;
 
     /**
-     * Timestamp at the last renewal of counter_hi field.
+     * The timestamp at the last renewal of counter_hi field.
      */
     private long tsCounterHi = 0;
 
     /**
-     * Status code reported at the last generation.
+     * The status code reported at the last generation.
      */
     private @NotNull Status lastStatus = Status.NOT_EXECUTED;
 
     /**
-     * Random number generator used by the generator.
+     * The random number generator used by the generator.
      */
     private final @NotNull Random random;
 
@@ -40,43 +62,89 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
     /**
      * Creates a generator object with a specified random number generator.
      *
-     * @param random random number generator (should be cryptographically strong and securely seeded).
+     * @param random A random number generator (which should be cryptographically strong and securely seeded).
      */
     public Scru128Generator(@NotNull Random random) {
         this.random = Objects.requireNonNull(random);
     }
 
     /**
-     * Generates a new SCRU128 ID object.
+     * Generates a new SCRU128 ID object from the current timestamp.
      * <p>
-     * This method is thread-safe; multiple threads can call it concurrently.
+     * See the {@link Scru128Generator} class documentation for the description.
      *
-     * @return new SCRU128 ID object.
+     * @return A new SCRU128 ID object.
      */
     public synchronized @NotNull Scru128Id generate() {
         return generateCore(System.currentTimeMillis());
     }
 
     /**
-     * Generates a new SCRU128 ID object with the timestamp passed.
+     * Experimental: Generates a new SCRU128 ID object from the current timestamp, guaranteeing the monotonic order
+     * of generated IDs despite a significant timestamp rollback.
      * <p>
-     * Unlike {@link #generate()}, this method is NOT thread-safe. The generator object should be protected from
+     * See the {@link Scru128Generator} class documentation for the description.
+     *
+     * @return A new SCRU128 ID object.
+     */
+    public synchronized @Nullable Scru128Id generateNoRewind() {
+        return generateCoreNoRewind(System.currentTimeMillis(), DEFAULT_ROLLBACK_ALLOWANCE);
+    }
+
+    /**
+     * Generates a new SCRU128 ID object from the timestamp passed.
+     * <p>
+     * See the {@link Scru128Generator} class documentation for the description.
+     * <p>
+     * Unlike {@link #generate}, this method is NOT thread-safe. The generator object should be protected from
      * concurrent accesses using a mutex or other synchronization mechanism to avoid race conditions.
      *
-     * @param timestamp 48-bit timestamp field value.
-     * @return new SCRU128 ID object.
-     * @throws IllegalArgumentException if the argument is not a 48-bit positive integer.
+     * @param timestamp A 48-bit timestamp field value.
+     * @return A new SCRU128 ID object.
+     * @throws IllegalArgumentException if the timestamp is not a 48-bit positive integer.
      */
     public @NotNull Scru128Id generateCore(long timestamp) {
+        long rollbackAllowance = DEFAULT_ROLLBACK_ALLOWANCE;
+        Scru128Id value = generateCoreNoRewind(timestamp, rollbackAllowance);
+        if (value == null) {
+            // reset state and resume
+            this.timestamp = 0;
+            tsCounterHi = 0;
+            value = generateCoreNoRewind(timestamp, rollbackAllowance);
+            lastStatus = Status.CLOCK_ROLLBACK;
+            assert value != null;
+        }
+        return value;
+    }
+
+    /**
+     * Experimental: Generates a new SCRU128 ID object from the timestamp passed, guaranteeing the monotonic order of
+     * generated IDs despite a significant timestamp rollback.
+     * <p>
+     * See the {@link Scru128Generator} class documentation for the description.
+     * <p>
+     * Unlike {@link #generateNoRewind}, this method is NOT thread-safe. The generator object should be protected from
+     * concurrent accesses using a mutex or other synchronization mechanism to avoid race conditions.
+     *
+     * @param timestamp         A 48-bit timestamp field value.
+     * @param rollbackAllowance The amount of timestamp rollback that is considered significant. A suggested value is
+     *                          {@code 10_000} (milliseconds).
+     * @return A new SCRU128 ID object.
+     * @throws IllegalArgumentException if the timestamp is not a 48-bit positive integer.
+     */
+    public @Nullable Scru128Id generateCoreNoRewind(long timestamp, long rollbackAllowance) {
         if (timestamp < 1 || timestamp > Scru128.MAX_TIMESTAMP) {
             throw new IllegalArgumentException("`timestamp` must be a 48-bit positive integer");
+        } else if (rollbackAllowance < 0 || rollbackAllowance > Scru128.MAX_TIMESTAMP) {
+            throw new IllegalArgumentException("`rollbackAllowance` out of reasonable range");
         }
 
-        lastStatus = Status.NEW_TIMESTAMP;
         if (timestamp > this.timestamp) {
             this.timestamp = timestamp;
             counterLo = random.nextInt() & Scru128.MAX_COUNTER_LO;
-        } else if (timestamp + 10000 > this.timestamp) {
+            lastStatus = Status.NEW_TIMESTAMP;
+        } else if (timestamp + rollbackAllowance > this.timestamp) {
+            // go on with previous timestamp if new one is not much smaller
             counterLo++;
             lastStatus = Status.COUNTER_LO_INC;
             if (counterLo > Scru128.MAX_COUNTER_LO) {
@@ -92,11 +160,8 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
                 }
             }
         } else {
-            // reset state if clock moves back by ten seconds or more
-            tsCounterHi = 0;
-            this.timestamp = timestamp;
-            counterLo = random.nextInt() & Scru128.MAX_COUNTER_LO;
-            lastStatus = Status.CLOCK_ROLLBACK;
+            // abort if clock moves back to unbearable extent
+            return null;
         }
 
         if (this.timestamp - tsCounterHi >= 1000 || tsCounterHi < 1) {
@@ -113,7 +178,7 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
      * Note that the generator object should be protected from concurrent accesses during the sequential calls to a
      * generation method and this method to avoid race conditions.
      *
-     * @return status code from the last generation of ID.
+     * @return A status code from the last generation of ID.
      */
     public @NotNull Status getLastStatus() {
         return lastStatus;
@@ -122,7 +187,7 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
     /**
      * Returns an infinite iterator object that produces a new ID for each call of {@code next()}.
      *
-     * @return infinite iterator.
+     * @return An infinite iterator.
      */
     @Override
     public @NotNull Iterator<@NotNull Scru128Id> iterator() {
@@ -142,9 +207,9 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
     /**
      * Returns a new SCRU128 ID object for each call, infinitely.
      * <p>
-     * This method is a synonym for {@link #generate()} to use {@code this} as an infinite iterator.
+     * This method is a synonym for {@link #generate} to use {@code this} as an infinite iterator.
      *
-     * @return new SCRU128 ID object.
+     * @return A new SCRU128 ID object.
      */
     @Override
     public @NotNull Scru128Id next() {
@@ -152,7 +217,7 @@ public class Scru128Generator implements Iterable<@NotNull Scru128Id>, Iterator<
     }
 
     /**
-     * Status code returned by {@link #getLastStatus()} method.
+     * The status code returned by {@link #getLastStatus} method.
      */
     public enum Status {
         /**